backend/requirements.txt
timm
torch
torchvision
matplotlib
tqdm
opencv-python
ultralytics
pillow
pydantic
psycopg2
fastapi
uvicorn
